+++ workspace/packages/apis/siyuan/tests/config.ts
@@ -1,0 +1,2 @@
+export const SERVER = 'http://localhost:6806';
+export const TOKEN = '07rea1i6pp6orqbu';