+++ workspace/packages/apis/siyuan/types/siyuan/getBlockBreadcrumb.d.ts
@@ -1,0 +1,20 @@
+// /api/block/getBlockBreadcrumb
+
+export interface IPayload {
+    id: string;
+    excludeTypes: BlockType[];
+}
+
+export interface Data {
+    id: ID;
+    name: string;
+    type: BlockType;
+    subType: BlockSubType;
+    children?: Data[];
+}
+
+export interface IResponse {
+    code: number;
+    msg: string;
+    data: Data[];
+}