+++ workspace/packages/apis/siyuan/types/siyuan/version.d.ts
@@ -1,0 +1,7 @@
+// /api/system/version
+
+export interface IResponse {
+    code: number;
+    msg: string;
+    data: string;
+}